+++ src/views/Transaction/Nomad/Details.vue
@@ -33,8 +33,8 @@
 
     <!-- link to block explorer -->
     <div class="flex flex-row">
-      <n-text>Block Explorer</n-text>
-      <a :href="explorerLink" target="_blank">
+      <a :href="explorerLink" class="flex items-center hover:underline" target="_blank">
+        <n-text>Block Explorer</n-text>
         <img
           src="@/assets/icons/arrow-right-up.svg"
           alt="open"