+++ CHANGELOG.md
@@ -6,3 +6,4 @@ This change log is manually updated at the moment.
 Unreleased
 --------------------------------
 - adds basic changelog
+- bunch of small tx details UI fixes